Refactor Atmospec submission step (#136)
Use the same base class as the OptimizationStep.
Does not refactor the UI yet,
but we now update UI from the loaded process parameters.
