+++ aiidalab_ispg/optimization_steps.py
@@ -11,9 +11,10 @@ from aiida.orm import load_code
 from aiida.plugins import DataFactory, WorkflowFactory
 from aiidalab_widgets_base import WizardAppWidgetStep
 
-from .input_widgets import MoleculeSettings, GroundStateSettings
+from .input_widgets import CodeSettings, MoleculeSettings, GroundStateSettings
 from .widgets import ResourceSelectionWidget
-from .steps import CodeSettings, SubmitWorkChainStepBase
+from .steps import SubmitWorkChainStepBase
+from .utils import MEMORY_PER_CPU
 
 try:
     from aiidalab_atmospec_workchain.optimization import ConformerOptimizationWorkChain
@@ -101,19 +102,22 @@ class SubmitOptimizationWorkChainStep(SubmitWorkChainStepBase):
 
     @traitlets.observe("process")
     def _observe_process(self, change):
-        self._update_state()
-        process = change["new"]
-        if process is None:
-            return
-        try:
-            parameters = process.base.extras.get("builder_parameters")
-            self._update_ui_from_parameters(OptimizationParameters(**parameters))
-        except AttributeError as e:
-            # extras do not exist or are incompatible, ignore this problem
-            pass
+        with self.hold_trait_notifications():
+            process = change["new"]
+            if process is not None:
+                self.input_structure = process.inputs.structure
+                try:
+                    parameters = process.base.extras.get("builder_parameters")
+                    self._update_ui_from_parameters(
+                        OptimizationParameters(**parameters)
+                    )
+                except (AttributeError, KeyError, TypeError):
+                    # extras do not exist or are incompatible, ignore this problem
+                    # TODO: Maybe display warning?
+                    pass
+            self._update_state()
 
     def submit(self, _=None):
-
         assert self.input_structure is not None
 
         parameters = self._get_parameters_from_ui()
@@ -155,11 +159,18 @@ class SubmitOptimizationWorkChainStep(SubmitWorkChainStepBase):
 
     def _build_orca_params(self, params: OptimizationParameters) -> dict:
         """Prepare dictionary of ORCA parameters, as required by aiida-orca plugin"""
+        # WARNING: Here we implicitly assume, that ORCA will automatically select
+        # equilibrium solvation for ground state optimization,
+        # and non-equilibrium solvation for single point excited state calculations.
+        # This should be the default, but it would be better to be explicit.
+        input_keywords = [params.basis, params.method, "Opt", "AnFreq"]
+        if params.solvent != "None":
+            input_keywords.append(f"CPCM({params.solvent})")
         return {
             "charge": params.charge,
             "multiplicity": params.multiplicity,
             "input_blocks": {
                 "scf": {"convergence": "tight", "ConvForced": "true"},
             },
-            "input_keywords": [params.basis, params.method, "Opt", "AnFreq"],
+            "input_keywords": input_keywords,
         }
